[FEAT/#102] Shop 가게 상세 조회 - 기존의 service 코드를 리팩토링
